+++ src/C++/FileStore.cpp
@@ -130,7 +130,7 @@ void FileStore::populateCache()
   {
     int num;
     long offset;
-    size_t size;
+    int size;
 
     while ( FILE_FSCANF( headerFile, "%d,%ld,%lu ", &num, &offset, &size ) == 3 )
       m_offsets[ num ] = std::make_pair( offset, size );
@@ -192,7 +192,7 @@ throw ( IOException )
   long offset = ftell( m_msgFile );
   if ( offset < 0 ) 
     throw IOException( "Unable to get file pointer position from " + m_msgFileName );
-  size_t size = msg.size();
+  int size = (int)msg.size();
 
   if ( fprintf( m_headerFile, "%d,%ld,%lu ", msgSeqNum, offset, size ) < 0 )
     throw IOException( "Unable to write to file " + m_headerFileName );
